requirements.txt
asttokens==3.0.0
attrs==25.3.0
biocommons.seqrepo==0.6.11
biopython==1.85
bioutils==0.6.1
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
coloredlogs==15.0.1
configparser==7.2.0
cssselect2==0.8.0
decorator==5.2.1
executing==2.2.0
fonttools==4.58.4
hgvs==1.5.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.1.7
narwhals==1.43.1
numpy==2.3.1
packaging==25.0
pandas==2.3.0
Parsley==1.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
plotly==6.1.2
prompt_toolkit==3.0.51
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
pyphen==0.17.2
pysam==0.23.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
wheel==0.45.1
yoyo-migrations==9.0.0
zipp==3.23.0
zopfli==0.2.3.post1
